web_api: remove dead code, dedup helpers, move detection FSM out of utils
Cleanup of utils/web_api + utils/prompt_generation after the WebAPITesting
merge. Behaviour-preserving; full test suite green (94 passed, 1 skipped).

Dead code removed:
- response_analyzer_with_llm: broken __main__ demo, MagicMock import, and
  set_purpose / print_results / replace_account (unused).
- pattern_matcher: __main__ demo.
- openapi_parser: _print_api_details (called a non-existent get_paths),
  get_protected_endpoints, get_refresh_endpoints.
- llm_handler: write-only _get_created_objects accessor (+ its test and the
  stale commented-out test_call_llm block).

Redundant abstractions collapsed:
- endpoint_categorizer.categorize_endpoints_with_query now backs both the
  parser's and the documentation engine's previously-identical
  categorize_endpoints.
- New utils/web_api/http_response.extract_status_code_and_message replaces the
  status-line regex duplicated in the analyzer and the OpenAPI handler.

Detection state machine moved into the use-case:
- New usecases/web_api/detection_response_handler.DetectionResponseHandler holds
  the ExploreStep exploration FSM (handle_response, adjust_path*, get_next_path,
  check_if_successful, common_endpoints, ...). The base utils ResponseHandler is
  now phase-agnostic only (evaluate_result, extract_key_elements_of_response,
  parse_http_status_line, set_response_analyzer). The documentation engine builds
  the subclass; the testing engine keeps the slim base.
- PromptGenerationHelper.get_hint now uses ExploreStep.INSTANCE/SUBRESOURCE/QUERY
  instead of the magic step-numbers 2/3/6.
